fix: Reorganizacion de archivos(hook, contanst,utils)
