+++ src/investments/contanst/actionsContanst/tradingPanel.contanst.ts
@@ -1,0 +1,1 @@
+export const quicksButton = [1, 5, 10, 50];
